// project 2/src/types/resume.ts
export interface ExperienceItem {
  title: string;
  location: string;
  period: string;
  responsibilities: string[];
}

export interface ProjectItem {
  title: string;
  description: string;
  link: string;
  color: string;
}

export interface EducationItem {
  degree: string;
  institution: string;
  period: string;
}